feat: apply Justin comment
