+++ core/integration/testdata/sdks/only-codegen/.gitignore
@@ -1,0 +1,4 @@
+/dagger.gen.go
+/internal/dagger
+/internal/querybuilder
+/internal/telemetry
