+++ core/integration/testdata/sdks/only-codegen/dagger.json
@@ -1,0 +1,7 @@
+{
+  "name": "only-codegen",
+  "engineVersion": "v0.18.9",
+  "sdk": {
+    "source": "go"
+  }
+}
