+++ core/integration/testdata/sdks/only-codegen/main.go
@@ -1,0 +1,16 @@
+package main
+
+import (
+	"context"
+	"dagger/only-codegen/internal/dagger"
+)
+
+type OnlyCodegen struct{}
+
+func (o *OnlyCodegen) Codegen(
+	ctx context.Context,
+	modSource *dagger.ModuleSource,
+	introspectionJSON *dagger.File,
+) (*dagger.GeneratedCode, error) {
+	return dag.GeneratedCode(dag.Directory().WithNewFile("hello.txt", "Hello, world!")), nil
+}
