+++ core/integration/testdata/sdks/only-runtime/.gitignore
@@ -1,0 +1,4 @@
+/dagger.gen.go
+/internal/dagger
+/internal/querybuilder
+/internal/telemetry
