+++ core/integration/testdata/sdks/only-runtime/dagger.json
@@ -1,0 +1,7 @@
+{
+  "name": "only-runtime",
+  "engineVersion": "v0.18.9",
+  "sdk": {
+    "source": "go"
+  }
+}
